+++ pkg/innerring/innerring.go
@@ -417,15 +417,6 @@ func New(ctx context.Context, log *zap.Logger, cfg *config.Config, errChan chan<
 
 	serveMetrics(server, cfg)
 
-	// TODO: remove deprecated 'fschain_autodeploy' config option in future release
-	if cfg.IsSet("fschain_autodeploy") {
-		log.Warn("configuration option 'fschain_autodeploy' is deprecated, use 'fschain.disable_autodeploy' with the opposite value instead")
-		if cfg.IsSet("fschain.disable_autodeploy") {
-			return nil, fmt.Errorf("'fschain_autodeploy' and 'fschain.disable_autodeploy' set simultaneously")
-		}
-		cfg.FSChain.DisableAutodeploy = !cfg.FSChainAutodeploy
-	}
-
 	var localWSClient *rpcclient.WSClient // set if isLocalConsensus only
 
 	// create FS chain client
@@ -581,14 +572,6 @@ func New(ctx context.Context, log *zap.Logger, cfg *config.Config, errChan chan<
 		return nil, err
 	}
 
-	// TODO: remove deprecated 'without_mainnet' config option in future release
-	if cfg.IsSet("without_mainnet") {
-		log.Warn("configuration option 'without_mainnet' is deprecated, use 'mainnet.enabled' with the reverted value instead")
-		if cfg.IsSet("mainnet.enabled") {
-			return nil, fmt.Errorf("'without_mainnet' and 'mainnet.enabled' set simultaneously")
-		}
-		cfg.Mainnet.Enabled = !cfg.WithoutMainnet
-	}
 	server.withoutMainNet = !cfg.Mainnet.Enabled
 
 	if server.withoutMainNet {
@@ -627,15 +610,6 @@ func New(ctx context.Context, log *zap.Logger, cfg *config.Config, errChan chan<
 	log.Info("notary support",
 		zap.Bool("mainchain_enabled", !server.mainNotaryConfig.disabled),
 	)
-
-	// TODO: remove deprecated 'contracts' config option in future release
-	if cfg.IsSet("contracts") {
-		log.Warn("configuration option 'contracts' is deprecated, use 'mainnet.contracts' with the same values instead")
-		if cfg.IsSet("mainnet.contracts") {
-			return nil, fmt.Errorf("'contracts' and 'mainnet.contracts' set simultaneously")
-		}
-		cfg.Mainnet.Contracts = cfg.Contracts
-	}
 
 	// get all script hashes of contracts
 	server.contracts, err = initContracts(ctx, log,
@@ -697,15 +671,6 @@ func New(ctx context.Context, log *zap.Logger, cfg *config.Config, errChan chan<
 		return nil, err
 	}
 
-	// TODO: remove deprecated 'fee.main_chain' config option in future release
-	if cfg.IsSet("fee.main_chain") {
-		log.Warn("configuration option 'fee.main_chain' is deprecated, use 'mainnet.extra_fee' with the same value instead")
-		if cfg.IsSet("mainnet.extra_fee") {
-			return nil, fmt.Errorf("'fee.main_chain' and 'mainnet.extra_fee' set simultaneously")
-		}
-		cfg.Mainnet.ExtraFee = cfg.Fee.MainChain
-	}
-
 	neofsCli, err := neofsClient.NewFromMorph(server.mainnetClient, server.contracts.neofs,
 		fixedn.Fixed8(cfg.Mainnet.ExtraFee), neofsClient.TryNotary(), neofsClient.AsAlphabet())
 	if err != nil {
@@ -738,15 +703,6 @@ func New(ctx context.Context, log *zap.Logger, cfg *config.Config, errChan chan<
 	}
 
 	var alphaSync event.Handler
-
-	// TODO: remove deprecated 'governance.disable' config option in future release
-	if cfg.IsSet("governance.disable") {
-		log.Warn("configuration option 'governance.disable' is deprecated, use 'mainnet.disable_governance_sync' with the same value instead")
-		if cfg.IsSet("mainnet.disable_governance_sync") {
-			return nil, fmt.Errorf("'governance.disable' and 'mainnet.disable_governance_sync' set simultaneously")
-		}
-		cfg.Mainnet.DisableGovernanceSync = cfg.Governance.Disable
-	}
 
 	if server.withoutMainNet || cfg.Mainnet.DisableGovernanceSync {
 		alphaSync = func(event.Event) {