+++ CHANGELOG.md
@@ -19,11 +19,19 @@ Changelog for NeoFS Node
 - Lock objects with API <2.18 are no longer accepted (#3672)
 
 ### Removed
+- Deprecated `fschain_autodeploy`, `without_mainnet`, `governance.disable`, `fee.main_chain` and `contracts` IR config options (#3716)
 
 ### Updated
 - `github.com/nspcc-dev/neofs-contract` module to `v0.25.2-0.20251124180339-40ec608b4893` (#3670)
 
 ### Updating from v0.50.2
+Please remove the following deprecated configuration options from IR config:
+- `fschain_autodeploy`
+- `without_mainnet`
+- `governance.disable`
+- `fee.main_chain`
+- `contracts`
+They were removed in this version and using them will lead to failure to start the node.
 
 ## [0.50.2] - 2025-11-27